src/requirements.txt
# Alpha Detector Module Dependencies
# Phase 1: Foundation

# Database
psycopg2-binary==2.9.10
supabase==2.18.1

# WebSocket
websockets==15.0.1

# Async support
asyncio==4.0.0
aiohttp==3.9.1

# Environment management
python-dotenv==1.1.1

# Data processing
pandas==2.2.0
numpy==1.26.4

# Configuration
PyYAML==6.0.2

# Logging
structlog==24.1.0

# LLM Integration
openai==1.107.2
requests==2.32.5

# System monitoring
psutil==5.9.8

# Testing
pytest==8.2.0
pytest-asyncio==0.23.5
